+++ packages/mcp/CHANGELOG.md
@@ -1,6 +1,22 @@
 # Changelog
 
 
+## 0.5.0 (2026-01-10)
+
+- Add support for `with_components` macro. ([#703](https://github.com/OpenZeppelin/contracts-wizard/pull/703))
+- Add AccessControlDefaultAdminRules. ([#698](https://github.com/OpenZeppelin/contracts-wizard/pull/698))
+
+- **Breaking changes**:
+  - Use OpenZeppelin Contracts for Cairo v3.0.0. ([#757](https://github.com/OpenZeppelin/contracts-wizard/pull/757))
+  - Use OpenZeppelin Contracts for Cairo v3.0.0-alpha.3. ([#688](https://github.com/OpenZeppelin/contracts-wizard/pull/688))
+  - Use OpenZeppelin Contracts for Cairo v3.0.0-alpha.2. ([#663](https://github.com/OpenZeppelin/contracts-wizard/pull/663))
+  - Use OpenZeppelin Contracts for Cairo v3.0.0-alpha.1. ([#638](https://github.com/OpenZeppelin/contracts-wizard/pull/638))
+  - Use OpenZeppelin Contracts for Cairo v3.0.0-alpha.0. ([#623](https://github.com/OpenZeppelin/contracts-wizard/pull/623))
+
+- Updated dependencies [[`53c042a`](https://github.com/OpenZeppelin/contracts-wizard/commit/53c042af852097bace639b47ec2fbfc813e38ac9)]:
+  - @openzeppelin/wizard-cairo@3.0.0
+  - @openzeppelin/wizard-common@0.4.0
+
 ## 0.4.6 (2025-12-12)
 
 - Updated dependencies [[`ce60539`](https://github.com/OpenZeppelin/contracts-wizard/commit/ce6053924898e3ecbc8f80e05502786ddbc5e559)]: